refactor(level3): 코드 정리
Operator 클래스 getOp() 대신 invoke()

+++ src/level3/MyScanner.java
@@ -13,7 +13,7 @@ class MyScanner {
         double num;
 
         try {
-            num = sc.nextInt();
+            num = sc.nextDouble();
             sc.nextLine(); // 버퍼 비우기
 
             return Optional.of(num);
